Backend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
colorama==0.4.3
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
Faker==4.0.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.5
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
wrapt==1.11.2
